rename extractor and move expander method extensions
